+++ packages/logger/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @graphql-portal/logger
 
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies [20dc51d]
+  - @graphql-portal/types@0.18.0
+
 ## 0.4.0
 
 ### Minor Changes